1. 차트 데이터 요청시 데이터 베이스에서 읽어 오게 변경 2. 세션 관리 방식 변경 3. 증권사 의존적인 코드 제외시킴

+++ SmServer/SmMarketManager.cpp
@@ -160,7 +160,7 @@ void SmMarketManager::SendSymbolListByCategory(std::string user_id)
 int SmMarketManager::GetTotalCategoryCount()
 {
 	int total = 0;
-	for (int i = 0; i < _MarketList.size(); ++i) {
+	for (size_t i = 0; i < _MarketList.size(); ++i) {
 		SmMarket* market = _MarketList[i];
 		total += market->GetCategoryList().size();
 	}
